requirements.txt
fastapi
uvicorn
faiss-cpu
pandas
websockets